pyproject.toml
[project]
name = "abs-audiobook-recommender"
version = "0.1.0"
description = "An audiobook recommendation system for items in your local ABS library using Google GenAI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.53.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
